Move ComponentSettings and subcomponents into its own folder
