+++ frontend/form-builder/src/components/FormSettings/ComponentSettings/index.ts
@@ -1,0 +1,1 @@
+export { ComponentSettings } from "./ComponentSettings"
